Simplify and split server.tsx
